requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.6
idna==3.4
image==1.5.33
packaging==23.0
Pillow==9.4.0
pytele==0.1.0
pyTelegramBotAPI==4.9.0
pytesseract==0.3.10
PyYAML==6.0
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.14
vk==3.0
